types4

+++ types2.py
@@ -1,5 +1,5 @@
 x = '47'
-y = int(x)
+y = float(x)
 
 print(f'x is {type(x)}')
 print(f'x is {x}')
